+++ kele.el
@@ -643,6 +643,25 @@ to complete.  Returned value may not be up to date."
             proxy-status
             (propertize ")" 'face 'completions-annotations))))
 
+(cl-defun kele--namespaces-complete (&optional context)
+  "Complete input for namespaces.
+
+If user does not have permission to list namespaces, simply
+prompt user for verbatim string.
+
+If CONTEXT is not provided, use the current context."
+  (let ((ctx (or context (kele-current-context-name))))
+    (completing-read
+     (format "Namespace (%s): " ctx)
+     (when (kele--can-i
+            :resource "namespaces"
+            :group "core"
+            :verb 'list
+            :context ctx)
+       (-cut kele--resources-complete <> <> <>
+             :cands (kele--get-namespaces ctx)
+             :category 'kele-namespace)))))
+
 (cl-defun kele--resources-complete (str pred action &key cands category)
   "Complete input for selection of resources.
 
@@ -659,12 +678,7 @@ as."
 (defun kele-namespace-switch-for-context (context namespace)
   "Switch to NAMESPACE for CONTEXT."
   (interactive (let ((context (completing-read "Context: " #'kele--contexts-complete)))
-                 (list context
-                       (completing-read (format "Namespace (%s): " context)
-                                        (-cut kele--resources-complete <> <> <>
-                                              :cands (kele--get-namespaces
-                                                      context)
-                                              :category 'kele-namespace)))))
+                 (list context (kele--namespaces-complete context))))
   (kele-kubectl-do "config" "set-context" context "--namespace" namespace))
 
 (transient-define-suffix kele-namespace-switch-for-current-context (namespace)
@@ -681,11 +695,7 @@ as."
                  (oref transient--prefix scope)
                (kele-current-context-name))))
      (list
-      (completing-read
-       (format "Namespace (%s): " ctx)
-       (-cut kele--resources-complete <> <> <>
-             :cands (kele--get-namespaces ctx)
-             :category 'kele-namespace)))))
+      (kele--namespaces-complete ctx))))
   (kele-namespace-switch-for-context
    (if (and transient--prefix
             (slot-boundp transient--prefix 'scope))
@@ -1080,6 +1090,7 @@ If CONTEXT is not provided, use the current context."
         (progn
           (setf (cdr (assoc 'items data)) filtered-items)
           data)
+      ;; FIXME: This might send false error if filtered-items ends up being empty
       (signal 'error (format "Failed to fetch %s/%s/%s" group version kind)))))
 
 (cl-defun kele--fetch-resource-names (group version kind &key namespace context)
@@ -1287,12 +1298,7 @@ Assumes that the current Transient prefix's :scope is an alist w/ `context' key.
   ;; value.  If not present (or the scope is not an alist or the scope is not
   ;; defined), default to current context.
   (if-let ((context (alist-get 'context (oref transient--prefix scope))))
-      (completing-read
-       prompt
-       (-cut kele--resources-complete <> <> <>
-             :cands (kele--get-namespaces context)
-             :category 'kele-namespace)
-       nil t initial-input history)
+      (kele--namespaces-complete context)
     (error "Unexpected nil context in `%s'" (oref transient--prefix command))))
 
 (defclass kele--transient-scope-mutator (transient-option)
@@ -1620,6 +1626,13 @@ resource type to query for.
 KIND should be the plural form of the kind's name, e.g. \"pods\"
 instead of \"pod.\""
   :key "g"
+  ;; TODO: Disable if user does not have permission to get the given resource
+  ;; :if (lambda ()
+  ;;       (kele--can-i :resource kind
+  ;;                    :group (-let (((group version) (kele--groupversion-split
+  ;;                                                    group-version)))
+  ;;                             group)
+  ;;                    :verb 'get))
   :description
   (lambda ()
     (--> (oref transient--prefix scope)
@@ -1782,7 +1795,7 @@ Similar to `kele-dispatch'."
                                :selected (string-equal ctx ctx-current)))
                      (kele-context-names))))))
 
-(cl-defun kele--mk-self-subject-access-review (&key resource group (verb 'get))
+(cl-defun kele--mk-self-subject-access-review (&key resource group (verb 'get) version)
   "Stub out a SelfSubjectAccessReview for GROUP, RESOURCE, and VERB.
 
 Return the resulting SelfSubjectAccessReview in alist form."
@@ -1790,6 +1803,7 @@ Return the resulting SelfSubjectAccessReview in alist form."
     (kind . "SelfSubjectAccessReview")
     (spec . ((resourceAttributes . ((group . ,group)
                                     (resource . ,resource)
+                                    (version . ,(or version "*"))
                                     (verb . ,(symbol-name verb))))))))